Non-recursive rapids evaluation
(cherry picked from commit e429d520b2d6ce2ffc6ed8c5d027cba4399fe0cd)
